Don't perform integer promotions on the operand to a cast; this simplifies the AST, and can matter in some rare cases involving casts to vector types. Patch by Enea Zaffanella.
llvm-svn: 79126

+++ clang/test/Sema/vector-cast.c
@@ -20,7 +20,7 @@ void f()
 type 't1' and scalar type 'char *'}}
   v1 = (t1)(long long)10;
   v1 = (t1)(short)10; // -expected-error {{invalid conversion between vector \
-type 't1' and integer type 'int' of different size}}
+type 't1' and integer type 'short' of different size}}
   
   long long r1 = (long long)v1;
   short r2 = (short)v1; // -expected-error {{invalid conversion between vector \
